Added lexer (lex function)
The tokenizer now includes extra metadata about the token which the lexer uses to discern the type of token being handled. The lex function does this beyond the work that is preemptively done by the tokenizer in assigning the type of token. For example, the tokenize function preemptively marks a token as being of type TOKEN_TYPE_STRING_LITERAL after reading an opening quote that is not escaped. Similarly, the tokenizer automatically marks symbols such as the left and right parentheses as TOKEN_TYPE_LEFT_PARENTHESIS and TOKEN_TYPE_RIGHT_PARENTHESIS respectively.